refactor nutritional values card styles
